Move 'Range' to the package it belongs in
